refactor RacingGameController, RacingGame
RacingGame과 RacingGameController 기능 교환
RacingGameController에서 View와 관계를 맺고 RacingGame을 서비스레이어로 변경
